Extract shared devtools infrastructure into devtools/lib
Version discovery/import-map resolution, the sandboxed iframe runner
protocol, the 2.x/3.x preview compat shim, and demo-preset access move
into devtools/lib so upcoming tools can reuse them. The benchmark is
refactored on top with no behavior change.

Assisted by Claude Code - Fable 5
